Migrate is_equivalent_structure to resonance.utils
